更新
マップ自動生成追加、音楽管理クラス追加、シーン遷移管理クラス追加、共通データクラス追加、

+++ aqua/src/utility/point/point.h
@@ -102,7 +102,7 @@ namespace aqua
          *
          *  @return     点クラス
          */
-        CPoint& operator*=(float scalar)
+        CPoint& operator*=(int scalar)
         {
             x *= scalar;
             y *= scalar;
@@ -117,7 +117,7 @@ namespace aqua
          *
          *  @return     点クラス
          */
-        CPoint& operator/=(float scalar)
+        CPoint& operator/=(int scalar)
         {
             x /= scalar;
             y /= scalar;
