+++ game/src/game/game_object/stage/stage_object/rock/object_rock.cpp
@@ -12,7 +12,7 @@ void CObjectRock::Initialize(aqua::CVector3 position)
 	m_Model.position.y -= 8.5f;
 	m_Model.scale = aqua::CVector3::ONE * 2.25f;
 	m_Model.initial_axis = aqua::CVector3(0.0f, 1.0f, 0.0f);
-	m_Model.initial_angles = aqua::DegToRad(aqua::Rand(360, 1));
+	m_Model.initial_angles = aqua::DegToRad((float)aqua::Rand(360, 1));
 	m_Model.axis = aqua::CVector3(1.0f, 0.0f, 0.0f);
 	m_Model.angles = aqua::DegToRad((float)std::pow(-1.0f, aqua::Rand(1, 0)) * 90.0f);
 }
